+++ gcs/CHANGES.md
@@ -9,6 +9,21 @@
         fs.gs.encryption.algorithm (not set by default)
         fs.gs.encryption.key (not set by default)
         fs.gs.encryption.key.hash (not set by default)
+1.  Added a new output stream type which can be used by setting:
+
+        fs.gs.outputstream.type=FLUSHABLE_COMPOSITE
+
+    The `FLUSHABLE_COMPOSITE` output stream type behaves similarly to the
+    `SYNCABLE_COMPOSITE` type, except it also supports `hflush()`, which uses
+    the same implementation with `hsync()` in the `SYNCABLE_COMPOSITE` output
+    stream type.
+1.  Added a new output stream parameter
+
+        fs.gs.outputstream.sync.min.interval.ms (default: 0)
+
+    to configure the minimum time interval (milliseconds) between consecutive
+    syncs. This is to avoid getting rate limited by GCS. Default is 0 - no wait
+    between syncs.
 
 ### 2.1.1 - 2020-03-11